Reference

VambBenchmarks.Bin — Type

Bin(name::AbstractString, sequences, targets)

Bins each represent a bin created by the binner. Conceptually, they are simply a set of Sequence with a name attached. Practically, every Bin is benchmarked against all Genomes and Clades of a given Reference, so each Bin stores data about its intersection with every genome/clade, e.g. its purity and recall.

Like Sources, Bins also have an assembly size for a given source. This is the number of base pairs in the source covered by any sequence in the Bin, which is always a subset of the Source's assembly size.

Benchmark statistics for a Bin/Genome can be done with either assemblies or _genomes as the ground truth.

True positives (TP) are defined as the sum of assembly sizes over all sources in the genome

False positives (FP) are the sum of length of sequences in the bin not mapping to the Genome

False negatives (FN) is either the genome assembly size or genome size minus TP.

For Bin/Clade pairs B/C, recall is the maximal recall of B/Ch for all children Ch of C. Precision is the sum of lengths of sequences mapping to any child of the clade divided by the sum of lengths of all sequences in the bin.

See also: Binning, Genome, Clade

Examples

julia> bin = first(binning.bins)
Bin "C1"
  Sequences: 2
  Breadth:   65
  Intersecting 1 genomes

julia> first(bin.sequences)
Sequence("s1", 25)

julia> f1(first(ref.genomes), bin)
0.625

VambBenchmarks.Binning — Type

Binning

A Binning represents a set of Bins benchmarked against a Reference. Binnings can be created given a set of Bins and a Reference, where the bins may potentially be loaded from a .tsv file. The field binning.recoverable_genomes shows the maximal number of recoverable genomes at given recall levels given perfect binning. The fields recovered_asms and recovered_genomes are used for benchmarking, these are normally output using the print_matrix function.

See also: print_matrix, Bin, Reference

Examples

julia> bins = gold_standard(ref);

julia> bins isa Binning
true

julia> VambBenchmarks.n_nc(binning)
0

VambBenchmarks.Clade — Type

Clade{Genome}(name::AbstractString, child::Union{Clade{Genome}, Genome})

A Clade represents any clade above Genome. Every Genome is expected to belong to the same number of clades, e.g. there may be exactly 7 levels of clades above every Genome. Clades always have at least one child (which is either a Genome or a Clade one rank lower), and a parent, unless it's the unique top clade from which all other clades and genomes descend from. The rank of a Genome is 0, clades that contain genomes have rank 1, and clades containing rank-1 clades have rank 2 etc. By default, zero-indexed ranks correspond to OTU, species, genus, family, order, class, phylum and domain.

Examples

julia> top_clade(ref)
Genus "F", 3 genomes
├─ Species "D", 2 genomes
│  ├─ Genome(gA)
│  └─ Genome(gB)
└─ Species "E", 1 genome
   └─ Genome(gC)

julia> top_clade(ref).children
2-element Vector{Clade{Genome}}:
 Species "D", 2 genomes
 Species "E", 1 genome

VambBenchmarks.FlagSet — Type

FlagSet <: AbstractSet{Flag}

Flags are compact sets of Flag associated to a Genome. You can construct them from an iterable of Flag, e.g. a 1-element tuple. FlagSet support most set operations efficiently.

Examples

julia> flags = FlagSet((Flags.organism, Flags.virus));

julia> Flags.virus in flags
true

julia> isdisjoint(flags, FlagSet((Flags.organism,)))
false

See also: Flag, Genome

VambBenchmarks.Flags.Flag — Type

Flag

A flag is a boolean associated to a Genome, stored in a Flags object. A flag may be e.g. Flag.organism, signaling that the genome is known to be an organism.

See also: FlagSet, Genome

VambBenchmarks.Genome — Type

Genome(name::AbstractString [flags::FlagSet])

Genomes represent individual target genomes (organisms, plasmids, viruses etc), analogous to lowest-level clade that can be reconstructed. Conceptually, Genomes contain one or more Sources, and to a single parent Clade. They are identified uniquely among genomes by their name.

A genome have a genome size, which is the sum of the length of all its sources. We consider this to be the true size of the biological genome (assuming its full sequence is contained in its sources), as well as an assembly size, which represent the sum of the assembly sizes of each source.

See also: Clade, Source, mrca

Examples

julia> gA, gB, gC = collect(ref.genomes);

julia> mrca(gA, gB)
Species "D", 2 genomes
├─ Genome(gA)
└─ Genome(gB)

VambBenchmarks.Reference — Type

Reference

A Reference contains the ground truth to benchmark against. Conceptually, it consists of the following parts:

A list of genomes, each with sources

The full taxonomic tree, as lists of clades

A list of sequences, each with a list of (source, span) to where it maps.

Normally, the types FlagSet Genome, Source, Clade and Sequence do not need to be constructed manually, but are constructed when the Reference is loaded from a JSON file.

Construct a Reference with open(i -> Reference(i), "path/to/ref.json").

Examples

julia> ngenomes(ref)
3

julia> nseqs(ref)
11

julia> first(ref.genomes) isa Genome
true

See also: subset, Genome, Clade

VambBenchmarks.Sequence — Type

Sequence(name::AbstractString, length::Integer)

Type that represents a binnable sequence. Sequences do not contain other information than their name and their length, and are identified by their name.

Examples

julia> Sequence("abc", 5)
Sequence("abc", 5)

julia> Sequence("abc", 5) == Sequence("abc", 9)
true

julia> Sequence("abc", 0)
ERROR: ArgumentError: Cannot instantiate an empty sequence

VambBenchmarks.Source — Type

Source{Genome}(g::Genome, name::AbstractString, length::Integer)

Sources are the "ground truth" sequences that the binning attempts to recreate. For example, the assembled contigs of the reference genome (typically full, closed circular contigs) as found in NCBI or elsewhere are each Sources. Many Genomes only contain a single Source namely its full assembled genome. Each Source has a single parent Genome, and a unique name which identifies it.

Sources have zero or more mapping Sequences, that each map to the Source at a given span given by a UnitRange{Int}.

Sources have an assembly size, which is the number of base pairs where any sequence map to.

VambBenchmarks.mrca — Method

mrca(a::Node, b::Node)::Node

Compute the most recent common ancestor (MRCA) of a and b.

VambBenchmarks.print_matrix — Method

print_matrix(::Binning; level=0, assembly=true)

Print the number of reconstructed assemblies or genomes at the given taxonomic level (rank). Level 0 corresponds to genomes, level 1 to species, etc. If assembly, print the number of reconstructed assemblies, else print the level of reconstructed genomes.

See also: Binning

VambBenchmarks.subset! — Method

subset!(ref::Reference, sequences::Function, genomes::Function)::Reference

Mutate ref in place, removing genomes and sequences. Keep only sequences S where sequences(S) returns true and genomes G for which genomes(G) returns true.

See also: subset, Reference

Examples

julia> ref
Reference
  Genomes:    3
  Sequences:  11
  Ranks:      3
  Seq length: 10
  Assembled:  66.8 %

julia> subset(ref; genomes=g -> Flags.organism in flags(g))
Reference
  Genomes:    2
  Sequences:  11
  Ranks:      3
  Seq length: 10
  Assembled:  91.3 %

julia> VambBenchmarks.subset(ref; sequences=s -> length(s) ≥ 25)
Reference
  Genomes:    3
  Sequences:  9
  Ranks:      3
  Seq length: 25
  Assembled:  61.1 %

VambBenchmarks.subset — Method

subset(ref::Reference; kwargs...)

Non-mutating copying version of subset.

See also: subset!